Kill some debugging related %lu warnings.
git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@3163 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1

+++ src/omxFitFunctionBA81.cpp
@@ -422,12 +422,12 @@ ba81ComputeEMFit(omxFitFunction* oo, int want, FitContext *fc)
 	}
 
 	if (excluded && estate->verbose >= 1) {
-		mxLog("%s: Hessian not positive definite for %lu/%lu items",
-		      oo->matrix->name, excluded, numItems);
+		mxLog("%s: Hessian not positive definite for %d/%d items",
+		      oo->matrix->name, (int) excluded, (int) numItems);
 	}
 	if (excluded == numItems) {
-		omxRaiseErrorf(globalState, "Hessian not positive definite for %lu/%lu items",
-			       excluded, numItems);
+		omxRaiseErrorf(globalState, "Hessian not positive definite for %d/%d items",
+			       (int) excluded, (int) numItems);
 	}
 
 	return Scale * ll;
